+++ addOns/automation/src/test/java/org/zaproxy/addon/automation/AutomationJobUnitTest.java
@@ -50,6 +50,7 @@ import org.zaproxy.addon.automation.tests.AbstractAutomationTest;
 import org.zaproxy.addon.automation.tests.AbstractAutomationTest.OnFail;
 import org.zaproxy.addon.automation.tests.AutomationAlertTest;
 import org.zaproxy.addon.automation.tests.AutomationStatisticTest;
+import org.zaproxy.addon.automation.tests.UrlPresenceTest;
 import org.zaproxy.zap.extension.alert.ExtensionAlert;
 import org.zaproxy.zap.extension.stats.ExtensionStats;
 import org.zaproxy.zap.extension.stats.InMemoryStats;
@@ -263,6 +264,39 @@ class AutomationJobUnitTest {
         assertThat(addedTest.getData().getName(), is(equalTo(name)));
         assertThat(addedTest.getData().getStatistic(), is(equalTo(statistic)));
         assertThat(addedTest.getData().getOperator(), is(equalTo(operator)));
+        assertThat(addedTest.getData().getOnFail(), is(equalTo(OnFail.WARN)));
+    }
+
+    @Test
+    void shouldAddUrlPresenceTests() {
+        // Given
+        TestParamContainer tpc = new TestParamContainer();
+        AutomationJob job = new AutomationJobImpl(tpc);
+        AutomationProgress progress = new AutomationProgress();
+        String name = "example name";
+        String type = "url";
+        String url = "http://example.com";
+        String onFail = "warn";
+        String operator = "or";
+
+        LinkedHashMap<String, Object> test = new LinkedHashMap<>();
+        test.put("name", name);
+        test.put("type", type);
+        test.put("url", url);
+        test.put("onFail", onFail);
+        test.put("operator", operator);
+        ArrayList<LinkedHashMap<String, Object>> tests = new ArrayList<>();
+        tests.add(test);
+
+        // When
+        job.addTests(tests, progress);
+
+        // Then
+        UrlPresenceTest addedTest = (UrlPresenceTest) job.getTests().get(0);
+        assertThat(progress.hasErrors(), is(equalTo(false)));
+        assertThat(progress.hasWarnings(), is(equalTo(false)));
+        assertThat(addedTest.getData().getName(), is(equalTo(name)));
+        assertThat(addedTest.getData().getUrl(), is(equalTo(url)));
         assertThat(addedTest.getData().getOnFail(), is(equalTo(OnFail.WARN)));
     }